+++ crates/fp/Cargo.toml
@@ -9,5 +9,4 @@ edition = "2018"
 [dependencies]
 
 [dev-dependencies]
-assert-impl = "0.1.3"
-rand = "0.8.3"
+rand = "0.8.4"
