--- crates/fps/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "fps"
-version = "0.1.0"
-authors = ["Kana <ngtkana@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-fp = { path = "../fp" }
-
-[dev-dependencies]
-itertools = "0.10.0"
-rand = "0.8.3"
